requirements.txt
Django==2.2.1
django-rest-framework==0.1.0
djangorestframework==3.9.4
pytz==2019.1
sqlparse==0.3.0
